// src/api/sqlmodule/model/finalizationReason.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type FinalizationReason = typeof FinalizationReason[keyof typeof FinalizationReason];


export const FinalizationReason = {
  Manual: 'Manual',
  PerfectScore: 'PerfectScore',
  AttemptsExhausted: 'AttemptsExhausted',
  Expired: 'Expired',
  EducationClosed: 'EducationClosed',
  Restarted: 'Restarted',
} as const;