// src/api/sqlmodule/model/hintGroup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type HintGroup = typeof HintGroup[keyof typeof HintGroup];


export const HintGroup = {
  Result: 'Result',
  RequiredConstructs: 'RequiredConstructs',
  ForbiddenConstructs: 'ForbiddenConstructs',
  RequiredTables: 'RequiredTables',
  ForbiddenTables: 'ForbiddenTables',
} as const;