// src/api/sqlmodule/model/progressStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type ProgressStatus = typeof ProgressStatus[keyof typeof ProgressStatus];


export const ProgressStatus = {
  Active: 'Active',
  Finalizing: 'Finalizing',
  CompletionPending: 'CompletionPending',
  CompletionFailed: 'CompletionFailed',
  Completed: 'Completed',
  Expired: 'Expired',
} as const;