// src/api/sqlmodule/model/sqlConstruct.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type SqlConstruct = typeof SqlConstruct[keyof typeof SqlConstruct];


export const SqlConstruct = {
  Join: 'Join',
  InnerJoin: 'InnerJoin',
  LeftJoin: 'LeftJoin',
  RightJoin: 'RightJoin',
  FullJoin: 'FullJoin',
  GroupBy: 'GroupBy',
  Having: 'Having',
  Distinct: 'Distinct',
  Subquery: 'Subquery',
  Cte: 'Cte',
  WindowFunction: 'WindowFunction',
} as const;