// src/api/sqlmodule/model/validationCheckKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type ValidationCheckKind = typeof ValidationCheckKind[keyof typeof ValidationCheckKind];


export const ValidationCheckKind = {
  MainDatasetResult: 'MainDatasetResult',
  RequiredConstruct: 'RequiredConstruct',
  ForbiddenConstruct: 'ForbiddenConstruct',
  RequiredTable: 'RequiredTable',
  ForbiddenTable: 'ForbiddenTable',
} as const;