// src/api/sqlmodule/model/validationCheckStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type ValidationCheckStatus = typeof ValidationCheckStatus[keyof typeof ValidationCheckStatus];


export const ValidationCheckStatus = {
  Passed: 'Passed',
  Failed: 'Failed',
  NotEvaluated: 'NotEvaluated',
} as const;