// src/api/sqlmodule/model/validationConfigurationState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type ValidationConfigurationState = typeof ValidationConfigurationState[keyof typeof ValidationConfigurationState];


export const ValidationConfigurationState = {
  Draft: 'Draft',
  Published: 'Published',
} as const;